Usar long para ids
git-svn-id: http://jira.uji.es/repos/uji/PAR/uji-par/trunk@16966 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-par-base/src/main/java/es/uji/apps/par/dao/ComprasDAO.java
@@ -72,7 +72,7 @@ public class ComprasDAO
     }
 
     @Transactional
-    public void marcarPagada(int idCompra)
+    public void marcarPagada(Long idCompra)
     {
         CompraDTO compra = getCompraById(idCompra);
         
@@ -82,7 +82,7 @@ public class ComprasDAO
     }
 
     @Transactional
-    public void borrarCompraNoPagada(Integer idCompra)
+    public void borrarCompraNoPagada(Long idCompra)
     {
         CompraDTO compra = getCompraById(idCompra);
